requirements.txt
pandas
numpy
openpyxl
pyarrow
pandera
python-dotenv
unidecode
jupyter
matplotlib